+++ src-tauri/src/audio/effects/mod.rs
@@ -5,26 +5,22 @@
 //! moves and mute toggles take effect within a couple of milliseconds without
 //! restarting the pipeline.
 
-use std::sync::atomic::{AtomicBool, AtomicU32, Ordering};
-use std::sync::Arc;
-
-use serde_json::Value;
-
-use crate::audio::graph::EffectSpec;
 use crate::audio::plugins::host_api::HostedEffect;
 
 /// Fixed DSP block size; hosted-plugin scratch buffers are sized to it. Must
 /// stay >= the pipeline's `DSP_BLOCK_FRAMES`, or a block would overrun them.
-const PLUGIN_MAX_BLOCK: usize = 1024;
+pub(crate) const PLUGIN_MAX_BLOCK: usize = 1024;
 
 pub mod biquad;
 pub mod channel_balance;
 pub mod compressor;
+pub mod controls;
 pub mod de_esser;
 pub mod declick;
 pub mod delay;
 pub mod eq;
 pub mod gain;
+pub mod instantiate;
 pub mod level_meter;
 pub mod limiter;
 pub mod lufs_meter;
@@ -32,12 +28,11 @@ pub mod mute;
 pub mod noise_gate;
 pub mod noise_suppressor;
 pub(crate) mod offload;
+pub mod registry;
 pub mod reverb;
 pub mod saturator;
-mod util;
+pub(crate) mod util;
 pub mod waveform;
-
-use util::{db_to_linear, num, store_f32};
 
 use channel_balance::ChannelBalanceEffect;
 use compressor::CompressorEffect;
@@ -51,19 +46,14 @@ use limiter::LimiterEffect;
 pub use lufs_meter::{LufsHandle, LufsMeterEffect};
 use mute::MuteEffect;
 use noise_gate::NoiseGateEffect;
-use noise_suppressor::{NoiseSuppressorControls, NoiseSuppressorEffect};
+use noise_suppressor::NoiseSuppressorEffect;
 use reverb::ReverbEffect;
 use saturator::SaturatorEffect;
 pub use waveform::{WaveformEffect, WaveformHandle};
 
-/// Shared atom for a dynamic-gain effect (compressor / noise gate / limiter).
-/// The audio thread writes the block's minimum gain each block; the meter tick
-/// thread reads it and emits `audio://gr` to the frontend.
-#[derive(Clone)]
-pub struct GrHandle {
-    pub node_id: String,
-    pub gr_lin: Arc<AtomicU32>,
-}
+pub use controls::EffectControl;
+pub use instantiate::instantiate_effect;
+pub use registry::{EffectBuild, EffectRegistry, GrHandle};
 
 pub trait Effect: Send {
     fn process(&mut self, samples: &mut [f32], frames: usize);
@@ -150,944 +140,3 @@ impl RuntimeEffect {
     }
 }
 
-#[derive(Clone)]
-pub enum EffectControl {
-    Gain {
-        linear: Arc<AtomicU32>,
-    },
-    Mute {
-        muted: Arc<AtomicBool>,
-    },
-    ChannelBalance {
-        left: Arc<AtomicU32>,
-        right: Arc<AtomicU32>,
-    },
-    Saturator {
-        ceiling: Arc<AtomicU32>,
-        drive: Arc<AtomicU32>,
-    },
-    Eq {
-        /// One gain atomic per ISO octave band; see EQ_FREQUENCIES_HZ for order.
-        gains: [Arc<AtomicU32>; 10],
-    },
-    Limiter {
-        ceiling: Arc<AtomicU32>,
-        release_ms: Arc<AtomicU32>,
-    },
-    Compressor {
-        threshold_db: Arc<AtomicU32>,
-        ratio: Arc<AtomicU32>,
-        attack_ms: Arc<AtomicU32>,
-        release_ms: Arc<AtomicU32>,
-        knee_db: Arc<AtomicU32>,
-        makeup_db: Arc<AtomicU32>,
-    },
-    NoiseGate {
-        threshold_db: Arc<AtomicU32>,
-        range_db: Arc<AtomicU32>,
-        attack_ms: Arc<AtomicU32>,
-        hold_ms: Arc<AtomicU32>,
-        release_ms: Arc<AtomicU32>,
-    },
-    Delay {
-        time_ms: Arc<AtomicU32>,
-        feedback: Arc<AtomicU32>,
-        mix: Arc<AtomicU32>,
-    },
-    Reverb {
-        room_size: Arc<AtomicU32>,
-        damping: Arc<AtomicU32>,
-        width: Arc<AtomicU32>,
-        mix: Arc<AtomicU32>,
-    },
-    NoiseSuppressor {
-        controls: NoiseSuppressorControls,
-    },
-    Declick {
-        sensitivity: Arc<AtomicU32>,
-        max_width_ms: Arc<AtomicU32>,
-    },
-    DeEsser {
-        frequency: Arc<AtomicU32>,
-        threshold_db: Arc<AtomicU32>,
-        ratio: Arc<AtomicU32>,
-    },
-    Plugin {
-        // Shared with the RT `PluginNode`; UI param writes flow through it.
-        events: Arc<crate::audio::plugins::ParamRing>,
-    },
-}
-
-impl EffectControl {
-    /// Unknown keys are silently ignored — the frontend pushes the full
-    /// camelCase payload of the node, only some keys map to live controls.
-    pub fn apply_update(&self, data: &Value) {
-        match self {
-            EffectControl::Gain { linear } => {
-                if let Some(db) = num(data, "gainDb") {
-                    store_f32(linear, db_to_linear(db));
-                }
-            }
-            EffectControl::Mute { muted } => {
-                if let Some(b) = data.get("muted").and_then(Value::as_bool) {
-                    muted.store(b, Ordering::Relaxed);
-                }
-            }
-            EffectControl::ChannelBalance { left, right } => {
-                if let Some(db) = num(data, "leftGainDb") {
-                    store_f32(left, db_to_linear(db));
-                }
-                if let Some(db) = num(data, "rightGainDb") {
-                    store_f32(right, db_to_linear(db));
-                }
-            }
-            EffectControl::Saturator { ceiling, drive } => {
-                if let Some(db) = num(data, "thresholdDb") {
-                    let c = db_to_linear(db).max(1e-6);
-                    store_f32(ceiling, c);
-                }
-                if let Some(db) = num(data, "driveDb") {
-                    store_f32(drive, db_to_linear(db));
-                }
-            }
-            EffectControl::Eq { gains } => {
-                if let Some(arr) = data.get("gainsDb").and_then(Value::as_array) {
-                    for (i, slot) in gains.iter().enumerate() {
-                        if let Some(v) = arr.get(i).and_then(Value::as_f64) {
-                            store_f32(slot, v as f32);
-                        }
-                    }
-                }
-            }
-            EffectControl::Limiter {
-                ceiling,
-                release_ms,
-            } => {
-                if let Some(db) = num(data, "ceilingDb") {
-                    store_f32(ceiling, db_to_linear(db).max(1e-6));
-                }
-                if let Some(ms) = num(data, "releaseMs") {
-                    store_f32(release_ms, ms.max(0.1));
-                }
-            }
-            EffectControl::Compressor {
-                threshold_db,
-                ratio,
-                attack_ms,
-                release_ms,
-                knee_db,
-                makeup_db,
-            } => {
-                if let Some(v) = num(data, "thresholdDb") {
-                    store_f32(threshold_db, v);
-                }
-                if let Some(v) = num(data, "ratio") {
-                    store_f32(ratio, v.max(1.0));
-                }
-                if let Some(v) = num(data, "attackMs") {
-                    store_f32(attack_ms, v.max(0.01));
-                }
-                if let Some(v) = num(data, "releaseMs") {
-                    store_f32(release_ms, v.max(0.1));
-                }
-                if let Some(v) = num(data, "kneeDb") {
-                    store_f32(knee_db, v.max(0.0));
-                }
-                if let Some(v) = num(data, "makeupDb") {
-                    store_f32(makeup_db, v);
-                }
-            }
-            EffectControl::NoiseGate {
-                threshold_db,
-                range_db,
-                attack_ms,
-                hold_ms,
-                release_ms,
-            } => {
-                if let Some(v) = num(data, "thresholdDb") {
-                    store_f32(threshold_db, v);
-                }
-                if let Some(v) = num(data, "rangeDb") {
-                    store_f32(range_db, v.min(0.0));
-                }
-                if let Some(v) = num(data, "attackMs") {
-                    store_f32(attack_ms, v.max(0.01));
-                }
-                if let Some(v) = num(data, "holdMs") {
-                    store_f32(hold_ms, v.max(0.0));
-                }
-                if let Some(v) = num(data, "releaseMs") {
-                    store_f32(release_ms, v.max(0.1));
-                }
-            }
-            EffectControl::Delay {
-                time_ms,
-                feedback,
-                mix,
-            } => {
-                if let Some(v) = num(data, "timeMs") {
-                    store_f32(time_ms, v.max(1.0));
-                }
-                if let Some(v) = num(data, "feedback") {
-                    store_f32(feedback, v.clamp(0.0, 0.95));
-                }
-                if let Some(v) = num(data, "mix") {
-                    store_f32(mix, v.clamp(0.0, 1.0));
-                }
-            }
-            EffectControl::Reverb {
-                room_size,
-                damping,
-                width,
-                mix,
-            } => {
-                if let Some(v) = num(data, "roomSize") {
-                    store_f32(room_size, v.clamp(0.0, 1.0));
-                }
-                if let Some(v) = num(data, "damping") {
-                    store_f32(damping, v.clamp(0.0, 1.0));
-                }
-                if let Some(v) = num(data, "width") {
-                    store_f32(width, v.clamp(0.0, 1.0));
-                }
-                if let Some(v) = num(data, "mix") {
-                    store_f32(mix, v.clamp(0.0, 1.0));
-                }
-            }
-            EffectControl::NoiseSuppressor { controls } => {
-                if let Some(v) = num(data, "attenuationLimitDb") {
-                    store_f32(&controls.atten_lim_db, v.max(0.0));
-                }
-                if let Some(v) = num(data, "postFilterBeta") {
-                    store_f32(&controls.pf_beta, v.max(0.0));
-                }
-                if let Some(v) = num(data, "minThreshDb") {
-                    store_f32(&controls.min_thresh_db, v);
-                }
-                if let Some(v) = num(data, "maxErbThreshDb") {
-                    store_f32(&controls.max_erb_thresh_db, v);
-                }
-                if let Some(v) = num(data, "maxDfThreshDb") {
-                    store_f32(&controls.max_df_thresh_db, v);
-                }
-            }
-            EffectControl::Declick {
-                sensitivity,
-                max_width_ms,
-            } => {
-                if let Some(v) = num(data, "sensitivity") {
-                    store_f32(sensitivity, v.clamp(0.0, 1.0));
-                }
-                if let Some(v) = num(data, "maxWidthMs") {
-                    store_f32(max_width_ms, v.clamp(0.3, 5.0));
-                }
-            }
-            EffectControl::DeEsser {
-                frequency,
-                threshold_db,
-                ratio,
-            } => {
-                if let Some(v) = num(data, "frequency") {
-                    store_f32(frequency, v.clamp(2000.0, 16000.0));
-                }
-                if let Some(v) = num(data, "thresholdDb") {
-                    store_f32(threshold_db, v.clamp(-80.0, 0.0));
-                }
-                if let Some(v) = num(data, "ratio") {
-                    store_f32(ratio, v.clamp(1.0, 12.0));
-                }
-            }
-            EffectControl::Plugin { events } => {
-                // `{ pluginParams: { "<paramId>": value } }` from the node UI.
-                if let Some(map) = data.get("pluginParams").and_then(Value::as_object) {
-                    for (id, v) in map {
-                        if let (Ok(id), Some(value)) = (id.parse::<u32>(), v.as_f64()) {
-                            events.push(id, value);
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-pub struct EffectBuild {
-    pub effect: RuntimeEffect,
-    /// Some only on the first instantiation per node id.
-    pub control: Option<EffectControl>,
-    /// Some only on the first instantiation per node id.
-    pub meter: Option<MeterHandle>,
-    /// Some only on the first instantiation per node id.
-    pub lufs: Option<LufsHandle>,
-    /// Some only on the first instantiation per node id, for GR-capable effects.
-    pub gr: Option<GrHandle>,
-    /// Some only on the first instantiation per node id, for oscilloscope nodes.
-    pub scope: Option<WaveformHandle>,
-    pub bypass: Arc<AtomicBool>,
-    pub bypass_is_new: bool,
-    /// The effect took the node's whole width, so the pipeline must hand it
-    /// every channel at once instead of splitting into stereo pairs.
-    pub full_width: bool,
-}
-
-/// Shared atomics keyed by node id so a fan-out effect (one node feeding
-/// multiple outputs) keeps live params in sync across instances.
-#[derive(Default)]
-pub struct EffectRegistry {
-    controls: std::collections::HashMap<String, EffectControl>,
-    bypasses: std::collections::HashMap<String, Arc<AtomicBool>>,
-    meters: std::collections::HashMap<String, MeterHandle>,
-    lufs: std::collections::HashMap<String, LufsHandle>,
-    gr_atomics: std::collections::HashMap<String, Arc<AtomicU32>>,
-    scopes: std::collections::HashMap<String, WaveformHandle>,
-    // Per-plugin UI->RT parameter queue, reused across rebuilds so the control
-    // handed to the frontend keeps reaching the current `PluginNode`.
-    plugin_param_rings: std::collections::HashMap<String, Arc<crate::audio::plugins::ParamRing>>,
-    // Plugin node ids that already claimed the editor-target (primary) instance
-    // in the current reconcile. A node feeding several real outputs is built
-    // once per output; only the first claim owns the editor, the rest are
-    // metering/duplicate instances parked in the graveyard.
-    plugin_primary_claimed: std::collections::HashSet<String>,
-}
-
-impl EffectRegistry {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Clears per-reconcile scratch. Call once before rebuilding the graphs so
-    /// the primary-instance claim is decided fresh each pass.
-    pub fn begin_reconcile(&mut self) {
-        self.plugin_primary_claimed.clear();
-    }
-}
-
-impl Drop for EffectRegistry {
-    /// A host keeps its own instance per plugin node so the UI thread can reach
-    /// it. Nothing else marks the end of a pipeline's life, so the registry
-    /// releases those holds as it goes; the instances themselves live on until
-    /// their RT nodes are dropped too.
-    fn drop(&mut self) {
-        for node_id in self.plugin_param_rings.keys() {
-            crate::audio::plugins::registry::forget(node_id);
-        }
-    }
-}
-
-pub fn instantiate_effect(
-    spec: &EffectSpec,
-    node_id: &str,
-    sample_rate: u32,
-    // False for file-recording outputs: an offline render outruns real time,
-    // so an expensive effect must process in place instead of on a worker.
-    realtime: bool,
-    // False when building the monitor graph: a plugin instantiated there is a
-    // metering-only duplicate, not the one its editor window attaches to.
-    primary: bool,
-    // Channels this node carries. An effect that can take them all says so
-    // through `EffectBuild::full_width`; the rest are driven one pair at a time.
-    channels: usize,
-    registry: &mut EffectRegistry,
-) -> EffectBuild {
-    let (bypass, bypass_is_new) = match registry.bypasses.get(node_id) {
-        Some(b) => (b.clone(), false),
-        None => {
-            let b = Arc::new(AtomicBool::new(spec.bypassed()));
-            registry.bypasses.insert(node_id.to_string(), b.clone());
-            (b, true)
-        }
-    };
-    let mk = |effect: RuntimeEffect,
-              control: Option<EffectControl>,
-              meter: Option<MeterHandle>,
-              lufs: Option<LufsHandle>,
-              gr: Option<GrHandle>,
-              scope: Option<WaveformHandle>| EffectBuild {
-        effect,
-        control,
-        meter,
-        lufs,
-        gr,
-        scope,
-        bypass: bypass.clone(),
-        bypass_is_new,
-        full_width: false,
-    };
-    match *spec {
-        EffectSpec::Gain(d) => match registry.controls.get(node_id) {
-            Some(EffectControl::Gain { linear }) => mk(
-                RuntimeEffect::Gain(GainEffect::from_state(linear.clone())),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ),
-            _ => {
-                let (e, c) = GainEffect::new(d);
-                registry.controls.insert(node_id.to_string(), c.clone());
-                mk(RuntimeEffect::Gain(e), Some(c), None, None, None, None)
-            }
-        },
-        EffectSpec::Mute(d) => match registry.controls.get(node_id) {
-            Some(EffectControl::Mute { muted }) => mk(
-                RuntimeEffect::Mute(MuteEffect::from_state(muted.clone())),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ),
-            _ => {
-                let (e, c) = MuteEffect::new(d);
-                registry.controls.insert(node_id.to_string(), c.clone());
-                mk(RuntimeEffect::Mute(e), Some(c), None, None, None, None)
-            }
-        },
-        EffectSpec::ChannelBalance(d) => match registry.controls.get(node_id) {
-            Some(EffectControl::ChannelBalance { left, right }) => mk(
-                RuntimeEffect::ChannelBalance(ChannelBalanceEffect::from_state(
-                    left.clone(),
-                    right.clone(),
-                )),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ),
-            _ => {
-                let (e, c) = ChannelBalanceEffect::new(d);
-                registry.controls.insert(node_id.to_string(), c.clone());
-                mk(
-                    RuntimeEffect::ChannelBalance(e),
-                    Some(c),
-                    None,
-                    None,
-                    None,
-                    None,
-                )
-            }
-        },
-        EffectSpec::Saturator(d) => match registry.controls.get(node_id) {
-            Some(EffectControl::Saturator { ceiling, drive }) => mk(
-                RuntimeEffect::Saturator(SaturatorEffect::from_state(
-                    ceiling.clone(),
-                    drive.clone(),
-                )),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ),
-            _ => {
-                let (e, c) = SaturatorEffect::new(d);
-                registry.controls.insert(node_id.to_string(), c.clone());
-                mk(RuntimeEffect::Saturator(e), Some(c), None, None, None, None)
-            }
-        },
-        EffectSpec::Eq(d) => match registry.controls.get(node_id) {
-            Some(EffectControl::Eq { gains }) => mk(
-                RuntimeEffect::Eq(EqEffect::from_state(gains.clone(), sample_rate)),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ),
-            _ => {
-                let (e, c) = EqEffect::new(d, sample_rate);
-                registry.controls.insert(node_id.to_string(), c.clone());
-                mk(RuntimeEffect::Eq(e), Some(c), None, None, None, None)
-            }
-        },
-        EffectSpec::LevelMeter(d) => match registry.meters.get(node_id) {
-            Some(handle) => mk(
-                RuntimeEffect::LevelMeter(LevelMeterEffect::from_handle(handle.clone())),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ),
-            None => {
-                let (e, handle) = LevelMeterEffect::new(d, node_id.to_string());
-                registry.meters.insert(node_id.to_string(), handle.clone());
-                mk(
-                    RuntimeEffect::LevelMeter(e),
-                    None,
-                    Some(handle),
-                    None,
-                    None,
-                    None,
-                )
-            }
-        },
-        EffectSpec::LufsMeter(d) => match registry.lufs.get(node_id) {
-            Some(handle) => mk(
-                RuntimeEffect::LufsMeter(LufsMeterEffect::from_handle(handle.clone(), sample_rate)),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ),
-            None => {
-                let (e, handle) = LufsMeterEffect::new(d, node_id.to_string(), sample_rate);
-                registry.lufs.insert(node_id.to_string(), handle.clone());
-                mk(
-                    RuntimeEffect::LufsMeter(e),
-                    None,
-                    None,
-                    Some(handle),
-                    None,
-                    None,
-                )
-            }
-        },
-        EffectSpec::Waveform(d) => match registry.scopes.get(node_id) {
-            Some(handle) => mk(
-                RuntimeEffect::Waveform(WaveformEffect::from_handle(handle.clone())),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ),
-            None => {
-                let (e, handle) = WaveformEffect::new(d, node_id.to_string(), sample_rate);
-                registry.scopes.insert(node_id.to_string(), handle.clone());
-                mk(
-                    RuntimeEffect::Waveform(e),
-                    None,
-                    None,
-                    None,
-                    None,
-                    Some(handle),
-                )
-            }
-        },
-        // Spectrum reuses the scope's time-domain capture and SCOPE_EVENT
-        // transport; the UI runs the FFT. No distinct runtime effect is needed.
-        EffectSpec::Spectrum(_) => match registry.scopes.get(node_id) {
-            Some(handle) => mk(
-                RuntimeEffect::Waveform(WaveformEffect::from_handle(handle.clone())),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ),
-            None => {
-                let (e, handle) = WaveformEffect::new_for(node_id.to_string(), sample_rate);
-                registry.scopes.insert(node_id.to_string(), handle.clone());
-                mk(
-                    RuntimeEffect::Waveform(e),
-                    None,
-                    None,
-                    None,
-                    None,
-                    Some(handle),
-                )
-            }
-        },
-        EffectSpec::Limiter(d) => match registry.controls.get(node_id) {
-            Some(EffectControl::Limiter {
-                ceiling,
-                release_ms,
-            }) => {
-                let lookahead_frames =
-                    ((d.lookahead_ms.max(0.1) * sample_rate as f32 / 1000.0) as usize).max(1);
-                let gr_arc = registry
-                    .gr_atomics
-                    .get(node_id)
-                    .cloned()
-                    .unwrap_or_else(|| Arc::new(AtomicU32::new(1.0f32.to_bits())));
-                registry
-                    .gr_atomics
-                    .insert(node_id.to_string(), gr_arc.clone());
-                // Republished on every rebuild: without this the meter thread
-                // loses the handle after the first reconcile and the readout
-                // freezes at its initial value.
-                let gr = GrHandle {
-                    node_id: node_id.to_string(),
-                    gr_lin: gr_arc.clone(),
-                };
-                mk(
-                    RuntimeEffect::Limiter(LimiterEffect::from_state(
-                        ceiling.clone(),
-                        release_ms.clone(),
-                        lookahead_frames,
-                        sample_rate,
-                        gr_arc,
-                    )),
-                    None,
-                    None,
-                    None,
-                    Some(gr),
-                    None,
-                )
-            }
-            _ => {
-                let (e, c, gr_arc) = LimiterEffect::new(d, sample_rate);
-                registry.controls.insert(node_id.to_string(), c.clone());
-                registry
-                    .gr_atomics
-                    .insert(node_id.to_string(), gr_arc.clone());
-                let gr = GrHandle {
-                    node_id: node_id.to_string(),
-                    gr_lin: gr_arc,
-                };
-                mk(
-                    RuntimeEffect::Limiter(e),
-                    Some(c),
-                    None,
-                    None,
-                    Some(gr),
-                    None,
-                )
-            }
-        },
-        EffectSpec::Compressor(d) => match registry.controls.get(node_id) {
-            Some(EffectControl::Compressor {
-                threshold_db,
-                ratio,
-                attack_ms,
-                release_ms,
-                knee_db,
-                makeup_db,
-            }) => {
-                let gr_arc = registry
-                    .gr_atomics
-                    .get(node_id)
-                    .cloned()
-                    .unwrap_or_else(|| Arc::new(AtomicU32::new(1.0f32.to_bits())));
-                registry
-                    .gr_atomics
-                    .insert(node_id.to_string(), gr_arc.clone());
-                // Republished on every rebuild: without this the meter thread
-                // loses the handle after the first reconcile and the readout
-                // freezes at its initial value.
-                let gr = GrHandle {
-                    node_id: node_id.to_string(),
-                    gr_lin: gr_arc.clone(),
-                };
-                mk(
-                    RuntimeEffect::Compressor(CompressorEffect::from_state(
-                        threshold_db.clone(),
-                        ratio.clone(),
-                        attack_ms.clone(),
-                        release_ms.clone(),
-                        knee_db.clone(),
-                        makeup_db.clone(),
-                        sample_rate,
-                        gr_arc,
-                    )),
-                    None,
-                    None,
-                    None,
-                    Some(gr),
-                    None,
-                )
-            }
-            _ => {
-                let (e, c, gr_arc) = CompressorEffect::new(d, sample_rate);
-                registry.controls.insert(node_id.to_string(), c.clone());
-                registry
-                    .gr_atomics
-                    .insert(node_id.to_string(), gr_arc.clone());
-                let gr = GrHandle {
-                    node_id: node_id.to_string(),
-                    gr_lin: gr_arc,
-                };
-                mk(
-                    RuntimeEffect::Compressor(e),
-                    Some(c),
-                    None,
-                    None,
-                    Some(gr),
-                    None,
-                )
-            }
-        },
-        EffectSpec::NoiseGate(d) => match registry.controls.get(node_id) {
-            Some(EffectControl::NoiseGate {
-                threshold_db,
-                range_db,
-                attack_ms,
-                hold_ms,
-                release_ms,
-            }) => {
-                let gr_arc = registry
-                    .gr_atomics
-                    .get(node_id)
-                    .cloned()
-                    .unwrap_or_else(|| Arc::new(AtomicU32::new(1.0f32.to_bits())));
-                registry
-                    .gr_atomics
-                    .insert(node_id.to_string(), gr_arc.clone());
-                // Republished on every rebuild: without this the meter thread
-                // loses the handle after the first reconcile and the readout
-                // freezes at its initial value.
-                let gr = GrHandle {
-                    node_id: node_id.to_string(),
-                    gr_lin: gr_arc.clone(),
-                };
-                mk(
-                    RuntimeEffect::NoiseGate(NoiseGateEffect::from_state(
-                        threshold_db.clone(),
-                        range_db.clone(),
-                        attack_ms.clone(),
-                        hold_ms.clone(),
-                        release_ms.clone(),
-                        sample_rate,
-                        gr_arc,
-                    )),
-                    None,
-                    None,
-                    None,
-                    Some(gr),
-                    None,
-                )
-            }
-            _ => {
-                let (e, c, gr_arc) = NoiseGateEffect::new(d, sample_rate);
-                registry.controls.insert(node_id.to_string(), c.clone());
-                registry
-                    .gr_atomics
-                    .insert(node_id.to_string(), gr_arc.clone());
-                let gr = GrHandle {
-                    node_id: node_id.to_string(),
-                    gr_lin: gr_arc,
-                };
-                mk(
-                    RuntimeEffect::NoiseGate(e),
-                    Some(c),
-                    None,
-                    None,
-                    Some(gr),
-                    None,
-                )
-            }
-        },
-        EffectSpec::Delay(d) => match registry.controls.get(node_id) {
-            Some(EffectControl::Delay {
-                time_ms,
-                feedback,
-                mix,
-            }) => mk(
-                RuntimeEffect::Delay(DelayEffect::from_state(
-                    time_ms.clone(),
-                    feedback.clone(),
-                    mix.clone(),
-                    sample_rate,
-                )),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ),
-            _ => {
-                let (e, c) = DelayEffect::new(d, sample_rate);
-                registry.controls.insert(node_id.to_string(), c.clone());
-                mk(RuntimeEffect::Delay(e), Some(c), None, None, None, None)
-            }
-        },
-        EffectSpec::Reverb(d) => match registry.controls.get(node_id) {
-            Some(EffectControl::Reverb {
-                room_size,
-                damping,
-                width,
-                mix,
-            }) => mk(
-                RuntimeEffect::Reverb(ReverbEffect::from_state(
-                    room_size.clone(),
-                    damping.clone(),
-                    width.clone(),
-                    mix.clone(),
-                    sample_rate,
-                )),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ),
-            _ => {
-                let (e, c) = ReverbEffect::new(d, sample_rate);
-                registry.controls.insert(node_id.to_string(), c.clone());
-                mk(RuntimeEffect::Reverb(e), Some(c), None, None, None, None)
-            }
-        },
-        EffectSpec::NoiseSuppressor(d) => match registry.controls.get(node_id) {
-            Some(EffectControl::NoiseSuppressor { controls }) => mk(
-                RuntimeEffect::NoiseSuppressor(NoiseSuppressorEffect::from_state(
-                    controls.clone(),
-                    sample_rate,
-                    realtime,
-                )),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ),
-            _ => {
-                let (e, c) = NoiseSuppressorEffect::new(d, sample_rate, realtime);
-                registry.controls.insert(node_id.to_string(), c.clone());
-                mk(
-                    RuntimeEffect::NoiseSuppressor(e),
-                    Some(c),
-                    None,
-                    None,
-                    None,
-                    None,
-                )
-            }
-        },
-        EffectSpec::Declick(d) => match registry.controls.get(node_id) {
-            Some(EffectControl::Declick {
-                sensitivity,
-                max_width_ms,
-            }) => mk(
-                RuntimeEffect::Declick(DeclickEffect::from_state(
-                    sensitivity.clone(),
-                    max_width_ms.clone(),
-                    sample_rate,
-                )),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ),
-            _ => {
-                let (e, c) = DeclickEffect::new(d, sample_rate);
-                registry.controls.insert(node_id.to_string(), c.clone());
-                mk(RuntimeEffect::Declick(e), Some(c), None, None, None, None)
-            }
-        },
-        EffectSpec::DeEsser(d) => match registry.controls.get(node_id) {
-            Some(EffectControl::DeEsser {
-                frequency,
-                threshold_db,
-                ratio,
-            }) => mk(
-                RuntimeEffect::DeEsser(DeEsserEffect::from_state(
-                    frequency.clone(),
-                    threshold_db.clone(),
-                    ratio.clone(),
-                    sample_rate,
-                )),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ),
-            _ => {
-                let (e, c) = DeEsserEffect::new(d, sample_rate);
-                registry.controls.insert(node_id.to_string(), c.clone());
-                mk(RuntimeEffect::DeEsser(e), Some(c), None, None, None, None)
-            }
-        },
-        EffectSpec::Plugin {
-            format,
-            ref path,
-            ref plugin_id,
-            ref state,
-            ..
-        } => {
-            // Empty path == node not yet configured: inert passthrough, not a
-            // failure. Silence is reserved for a real load error below.
-            if path.is_empty() {
-                crate::audio::plugins::registry::forget(node_id);
-                let muted = Arc::new(AtomicBool::new(false));
-                return mk(
-                    RuntimeEffect::Mute(MuteEffect::from_state(muted)),
-                    None,
-                    None,
-                    None,
-                    None,
-                    None,
-                );
-            }
-            // Surfaced as a load failure rather than guessed at: a path without
-            // a format is stored data we cannot act on.
-            let Some(format) = format else {
-                tracing::error!(node_id, path, "plugin node has no format");
-                let muted = Arc::new(AtomicBool::new(true));
-                return mk(
-                    RuntimeEffect::Mute(MuteEffect::from_state(muted)),
-                    None,
-                    None,
-                    None,
-                    None,
-                    None,
-                );
-            };
-            // Claimed only once the instance actually exists (below): a burned
-            // claim would leave the node with no editor target while the
-            // previous plugin stayed installed behind it.
-            let primary = primary && !registry.plugin_primary_claimed.contains(node_id);
-            // Every stereo pair shares the persistent per-node broadcast ring
-            // (kept across rebuilds); each pair reads it through its own cursor,
-            // so a UI write reaches all pairs, not just the first.
-            let ring = registry
-                .plugin_param_rings
-                .entry(node_id.to_string())
-                .or_insert_with(|| Arc::new(crate::audio::plugins::ParamRing::new()))
-                .clone();
-            let request = crate::audio::plugins::host_api::ActivateRequest {
-                node_id,
-                path,
-                plugin_id,
-                sample_rate,
-                max_frames: PLUGIN_MAX_BLOCK,
-                channels,
-                state: state.as_deref(),
-                primary,
-                params: ring.clone(),
-            };
-            match crate::audio::plugins::registry::activate(format, request) {
-                // Only the editor-target build publishes the control, so the UI
-                // writes reach the audible instance.
-                Ok(node) => {
-                    if primary {
-                        registry.plugin_primary_claimed.insert(node_id.to_string());
-                    }
-                    let control = primary.then_some(EffectControl::Plugin { events: ring });
-                    // The plugin took the node whole, so the pipeline must stop
-                    // splitting it into pairs and hand it every channel.
-                    let full_width = node.channels() == channels;
-                    let mut build = mk(
-                        RuntimeEffect::HostedPlugin(HostedEffect::new(node, realtime)),
-                        control,
-                        None,
-                        None,
-                        None,
-                        None,
-                    );
-                    build.full_width = full_width;
-                    build
-                }
-                Err(e) => {
-                    // Surface as silence, never a passthrough that hides the failure.
-                    tracing::error!(node_id, path, plugin_id, error = %e, "plugin failed to load");
-                    crate::audio::plugins::registry::forget(node_id);
-                    let muted = Arc::new(AtomicBool::new(true));
-                    mk(
-                        RuntimeEffect::Mute(MuteEffect::from_state(muted)),
-                        None,
-                        None,
-                        None,
-                        None,
-                        None,
-                    )
-                }
-            }
-        }
-    }
-}
